pyproject.toml
[project]
name = "research"
version = "0.1.0"
description = ""
authors = [
    {name = "asato425",email = "suasa425@icloud.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0" 
dependencies = [
    "python-dotenv (==1.1.1)",
    "langchain (==0.3.26)",
    "aiohappyeyeballs (==2.6.1)",
    "aiohttp (==3.12.14)",
    "aiosignal (==1.4.0)",
    "annotated-types (==0.7.0)",
    "anthropic (==0.60.0)",
    "anyio (==4.9.0)",
    "appdirs (==1.4.4)",
    "asttokens (==3.0.0)",
    "attrs (==25.3.0)",
    "backoff (==2.2.1)",
    "bcrypt (==4.3.0)",
    "build (==1.3.0)",
    "cachetools (==5.5.2)",
    "certifi (==2025.7.14)",
    "charset-normalizer (==3.4.2)",
    "chromadb (==1.0.16)",
    "click (==8.2.1)",
    "coloredlogs (==15.0.1)",
    "dataclasses-json (==0.6.7)",
    "datasets (==4.0.0)",
    "decorator (==5.2.1)",
    "dill (==0.3.8)",
    "diskcache (==5.6.3)",
    "distro (==1.9.0)",
    "durationpy (==0.10)",
    "executing (==2.2.0)",
    "fastapi (==0.116.1)",
    "filelock (==3.18.0)",
    "filetype (==1.2.0)",
    "flatbuffers (==25.2.10)",
    "frozenlist (==1.7.0)",
    "fsspec (==2025.3.0)",
    "gitdb (==4.0.12)",
    "gitpython (==3.1.45)",
    "google-ai-generativelanguage (==0.6.15)",
    "google-api-core (==2.25.1)",
    "google-api-python-client (==2.178.0)",
    "google-auth (==2.40.3)",
    "google-auth-httplib2 (==0.2.0)",
    "google-generativeai (==0.8.5)",
    "google-search-results (==2.4.2)",
    "googleapis-common-protos (==1.70.0)",
    "grandalf (==0.8)",
    "gritql (==0.2.0)",
    "grpcio (==1.74.0)",
    "grpcio-status (==1.71.2)",
    "grpcio-tools (==1.71.2)",
    "h11 (==0.16.0)",
    "hf-xet (==1.1.7)",
    "httpcore (==1.0.9)",
    "httplib2 (==0.22.0)",
    "httptools (==0.6.4)",
    "httpx (==0.28.1)",
    "httpx-sse (==0.4.1)",
    "huggingface-hub (==0.34.3)",
    "humanfriendly (==10.0)",
    "idna (==3.10)",
    "importlib-metadata (==8.7.0)",
    "importlib-resources (==6.5.2)",
    "ipython (==9.4.0)",
    "ipython-pygments-lexers (==1.1.1)",
    "jedi (==0.19.2)",
    "jiter (==0.10.0)",
    "jsonpatch (==1.33)",
    "jsonpointer (==3.0.0)",
    "jsonschema (==4.25.0)",
    "jsonschema-specifications (==2025.4.1)",
    "kubernetes (==33.1.0)",
    "langchain-anthropic (==0.3.18)",
    "langchain-chroma (==0.2.5)",
    "langchain-cli (==0.0.36)",
    "langchain-community (==0.3.27)",
    "langchain-core (==0.3.74)",
    "langchain-google-genai (==2.0.10)",
    "langchain-openai (==0.3.29)",
    "langchain-text-splitters (==0.3.8)",
    "langgraph (==0.5.4)",
    "langgraph-checkpoint (==2.1.1)",
    "langgraph-prebuilt (==0.5.2)",
    "langgraph-sdk (==0.1.74)",
    "langserve (==0.3.1)",
    "langsmith (==0.4.13)",
    "markdown-it-py (==3.0.0)",
    "marshmallow (==3.26.1)",
    "matplotlib-inline (==0.1.7)",
    "mdurl (==0.1.2)",
    "mmh3 (==5.2.0)",
    "mpmath (==1.3.0)",
    "multidict (==6.6.3)",
    "multiprocess (==0.70.16)",
    "mypy-extensions (==1.1.0)",
    "nest-asyncio (==1.6.0)",
    "numexpr (==2.11.0)",
    "numpy (==2.3.2)",
    "oauthlib (==3.3.1)",
    "onnxruntime (==1.22.1)",
    "openai (==1.97.1)",
    "opentelemetry-api (==1.36.0)",
    "opentelemetry-exporter-otlp-proto-common (==1.36.0)",
    "opentelemetry-exporter-otlp-proto-grpc (==1.36.0)",
    "opentelemetry-proto (==1.36.0)",
    "opentelemetry-sdk (==1.36.0)",
    "opentelemetry-semantic-conventions (==0.57b0)",
    "orjson (==3.11.0)",
    "ormsgpack (==1.10.0)",
    "overrides (==7.7.0)",
    "packaging (==24.2)",
    "pandas (==2.3.1)",
    "parso (==0.8.4)",
    "pexpect (==4.9.0)",
    "pillow (==11.3.0)",
    "posthog (==5.4.0)",
    "prompt-toolkit (==3.0.51)",
    "propcache (==0.3.2)",
    "proto-plus (==1.26.1)",
    "protobuf (==5.29.5)",
    "ptyprocess (==0.7.0)",
    "pure-eval (==0.2.3)",
    "pyarrow (==21.0.0)",
    "pyasn1 (==0.6.1)",
    "pyasn1-modules (==0.4.2)",
    "pybase64 (==1.4.2)",
    "pydantic (==2.11.7)",
    "pydantic-settings (==2.10.1)",
    "pydantic-core (==2.33.2)",
    "pygments (==2.19.2)",
    "pyparsing (==3.2.3)",
    "pypika (==0.48.9)",
    "pyproject-hooks (==1.2.0)",
    "pysbd (==0.3.4)",
    "python-dateutil (==2.9.0.post0)",
    "pytz (==2025.2)",
    "pyyaml (==6.0.2)",
    "ragas (==0.3.1)",
    "rapidfuzz (==3.13.0)",
    "referencing (==0.36.2)",
    "regex (==2024.11.6)",
    "requests (==2.32.4)",
    "requests-oauthlib (==2.0.0)",
    "requests-toolbelt (==1.0.0)",
    "rich (==14.0.0)",
    "rpds-py (==0.27.0)",
    "rsa (==4.9.1)",
    "setuptools (==80.9.0)",
    "shellingham (==1.5.4)",
    "six (==1.17.0)",
    "smmap (==5.0.2)",
    "sniffio (==1.3.1)",
    "sqlalchemy (==2.0.41)",
    "sse-starlette (==1.8.2)",
    "stack-data (==0.6.3)",
    "starlette (==0.47.2)",
    "sympy (==1.14.0)",
    "tavily-python (==0.7.10)",
    "tenacity (==8.5.0)",
    "tiktoken (==0.9.0)",
    "tokenizers (==0.21.4)",
    "tomlkit (==0.13.3)",
    "tqdm (==4.67.1)",
    "traitlets (==5.14.3)",
    "typer (==0.16.0)",
    "typing-inspect (==0.9.0)",
    "typing-inspection (==0.4.1)",
    "typing-extensions (==4.14.1)",
    "tzdata (==2025.2)",
    "uritemplate (==4.2.0)",
    "urllib3 (==2.5.0)",
    "uvicorn (==0.35.0)",
    "uvloop (==0.21.0)",
    "watchfiles (==1.1.0)",
    "wcwidth (==0.2.13)",
    "websocket-client (==1.8.0)",
    "websockets (==15.0.1)",
    "xxhash (==3.5.0)",
    "yarl (==1.20.1)",
    "zipp (==3.23.0)",
    "zstandard (==0.23.0)",
    "pygithub (>=2.7.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "research", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
